refactor(library/Attempt): moves "tryCatchStatements/safeTryCatch.ts" to "tryCatchStatements/safeAsync.ts"
